+++ dorm_annotations/example/lib/school.dart
@@ -19,9 +19,9 @@ abstract class _SchoolAddress {
   int get number;
 }
 
-@Model(name: 'escola', as: #schools, primaryKeyGenerator: _School._generate)
+@Model(name: 'escola', as: #schools)
 abstract class _School {
-  static String _generate(_School school, String id) => school.name;
+  static String $dorm$generateId(_School school, String id) => school.name;
 
   @Field(name: 'nome')
   String get name;